# swagger-spec.json
{"openapi":"3.0.0","paths":{"/clients":{"post":{"operationId":"ClientsController_create","summary":"","description":"Method to create a new client\n\n\nVisit my GitHub profile to see the code:\n[@ruslanguns](https://github.com/ruslanguns)","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateClientDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Clients"]},"get":{"operationId":"ClientsController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Clients"]}},"/clients/{id}":{"get":{"operationId":"ClientsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Clients"]},"put":{"operationId":"ClientsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateClientDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Clients"]},"delete":{"operationId":"ClientsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["Clients"]}},"/suppliers":{"post":{"operationId":"SuppliersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSupplierDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Supplier"]},"get":{"operationId":"SuppliersController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Supplier"]}},"/suppliers/{id}":{"get":{"operationId":"SuppliersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Supplier"]},"put":{"operationId":"SuppliersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSupplierDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Supplier"]},"delete":{"operationId":"SuppliersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["Supplier"]}},"/products":{"post":{"operationId":"ProductsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Products"]},"get":{"operationId":"ProductsController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Products"]}},"/products/{id}":{"get":{"operationId":"ProductsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Products"]},"put":{"operationId":"ProductsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Products"]},"delete":{"operationId":"ProductsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["Products"]}},"/purchase-orders":{"post":{"operationId":"PurchaseOrdersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePurchaseOrderDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Purchase orders"]},"get":{"operationId":"PurchaseOrdersController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Purchase orders"]}},"/purchase-orders/{id}":{"get":{"operationId":"PurchaseOrdersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Purchase orders"]},"put":{"operationId":"PurchaseOrdersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePurchaseOrderDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Purchase orders"]},"delete":{"operationId":"PurchaseOrdersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["Purchase orders"]}},"/sales-orders":{"post":{"operationId":"SalesOrdersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSalesOrderDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Sales orders"]},"get":{"operationId":"SalesOrdersController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["Sales orders"]}},"/sales-orders/{id}":{"get":{"operationId":"SalesOrdersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Sales orders"]},"patch":{"operationId":"SalesOrdersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSalesOrderDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Sales orders"]},"delete":{"operationId":"SalesOrdersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["Sales orders"]}}},"info":{"title":"Inventory SYS","description":"Creado con ♥ por [@ruslangonzalez](https://twitter.com/ruslangonzalez)","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateClientDto":{"type":"object","properties":{"idNumber":{"type":"string","minLength":4,"maxLength":20},"idType":{"type":"string"},"name":{"type":"string"}},"required":["idNumber","idType","name"]},"UpdateClientDto":{"type":"object","properties":{"idNumber":{"type":"string","minLength":4,"maxLength":20},"idType":{"type":"string"},"name":{"type":"string"}}},"CreateSupplierDto":{"type":"object","properties":{"idNumber":{"type":"string","minLength":4,"maxLength":20},"idType":{"type":"string"},"name":{"type":"string"}},"required":["idNumber","idType","name"]},"UpdateSupplierDto":{"type":"object","properties":{"idNumber":{"type":"string","minLength":4,"maxLength":20},"idType":{"type":"string"},"name":{"type":"string"}}},"CreateProductDto":{"type":"object","properties":{"sku":{"type":"string"},"sn":{"type":"string"},"stock":{"type":"number","minimum":0}},"required":["sku","sn","stock"]},"UpdateProductDto":{"type":"object","properties":{"sku":{"type":"string"},"sn":{"type":"string"},"stock":{"type":"number","minimum":0}}},"CreatePurchaseOrderDto":{"type":"object","properties":{"qty":{"type":"number","minimum":1},"supplierId":{"type":"string"},"productId":{"type":"string"}},"required":["qty","supplierId","productId"]},"UpdatePurchaseOrderDto":{"type":"object","properties":{"qty":{"type":"number","minimum":1},"supplierId":{"type":"string"},"productId":{"type":"string"}}},"CreateSalesOrderDto":{"type":"object","properties":{"qty":{"type":"number","minimum":1},"clientId":{"type":"string"},"productId":{"type":"string"}},"required":["qty","clientId","productId"]},"UpdateSalesOrderDto":{"type":"object","properties":{"qty":{"type":"number","minimum":1},"clientId":{"type":"string"},"productId":{"type":"string"}}}}}}